[FIX] `gpt` routes
- modified `gpt` routes. (from `kogpt`)
- need to engineering the prompt only
